requirements.txt
tensorflow-gpu==1.15
scipy
sklearn
matplotlib 
natsort
keras_applications>=1.0.7
image-classifiers==1.0.0
efficientnet==1.0.0
segmentation-models==1.0.1
albumentations==0.3.0
h5py==2.10.0